+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/configs/BzClientConfigs.java
@@ -4,6 +4,7 @@ public class BzClientConfigs {
     public static boolean playWrathOfHiveEffectMusic = true;
     public static boolean playSempiternalSanctumMusic = true;
     public static boolean renderBeeQueenBonusTradeItem = true;
+    public static boolean disableEssenceBlockShaders = false;
     public static boolean knowingEssenceHighlightBosses = true;
     public static boolean knowingEssenceHighlightMonsters = true;
     public static boolean knowingEssenceHighlightTamed = true;