+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/configs/BzWorldgenConfigs.java
@@ -3,6 +3,7 @@ package com.telepathicgrunt.the_bumblezone.configs;
 public class BzWorldgenConfigs {
 
     public static int beeDungeonRarity = 1;
+    public static int treeDungeonRarity = 2;
     public static int spiderInfestedBeeDungeonRarity = 8;
     public static double spawnerRateSpiderBeeDungeon = 0.2D;