+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/entities/mobs/RootminEntity.java
@@ -491,7 +491,7 @@ public class RootminEntity extends PathfinderMob implements Enemy {
                       !FlowerHeadwearHelmet.getFlowerHeadwear(player).isEmpty())) {
          BlockState blockState = blockItem.getBlock().defaultBlockState();
 
-         if (blockState.is(BzTags.ROOTMIN_ALLOWED_FLOWER) && !blockState.is(BzTags.ROOTMIN_FORCED_DISALLOWED_FLOWER)) {
+         if (blockState.is(BzTags.ROOTMIN_ALLOWED_FLOWERS) && !blockState.is(BzTags.ROOTMIN_FORCED_DISALLOWED_FLOWERS)) {
             if (!this.level().isClientSide()) {
                if (!instantBuild && this.getFlowerBlock() != null) {
                   ItemStack itemStack = new ItemStack(Items.DIAMOND_PICKAXE);