+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/modinit/BzTags.java
@@ -60,8 +60,8 @@ public class BzTags {
     public static final TagKey<Block> SENTRY_WATCHER_ALWAYS_DESTROY = TagKey.create(Registries.BLOCK, new ResourceLocation(Bumblezone.MODID, "sentry_watcher/always_destroy"));
     public static final TagKey<Block> SENTRY_WATCHER_FORCED_NEVER_DESTROY = TagKey.create(Registries.BLOCK, new ResourceLocation(Bumblezone.MODID, "sentry_watcher/forced_never_destroy"));
     public static final TagKey<Block> ROOTMIN_DEFAULT_FLOWERS = TagKey.create(Registries.BLOCK, new ResourceLocation(Bumblezone.MODID, "rootmin/default_flowers"));
-    public static final TagKey<Block> ROOTMIN_ALLOWED_FLOWER = TagKey.create(Registries.BLOCK, new ResourceLocation(Bumblezone.MODID, "rootmin/allowed_flower"));
-    public static final TagKey<Block> ROOTMIN_FORCED_DISALLOWED_FLOWER = TagKey.create(Registries.BLOCK, new ResourceLocation(Bumblezone.MODID, "rootmin/forced_disallowed_flower"));
+    public static final TagKey<Block> ROOTMIN_ALLOWED_FLOWERS = TagKey.create(Registries.BLOCK, new ResourceLocation(Bumblezone.MODID, "rootmin/allowed_flowers"));
+    public static final TagKey<Block> ROOTMIN_FORCED_DISALLOWED_FLOWERS = TagKey.create(Registries.BLOCK, new ResourceLocation(Bumblezone.MODID, "rootmin/forced_disallowed_flowers"));
     public static final TagKey<Block> ESSENCE_BLOCKS = TagKey.create(Registries.BLOCK, new ResourceLocation(Bumblezone.MODID, "essence_blocks"));
     public static final TagKey<Block> AIR_LIKE = TagKey.create(Registries.BLOCK, new ResourceLocation(Bumblezone.MODID, "air_like"));