+++ common/src/main/java/com/telepathicgrunt/the_bumblezone/worldgen/features/TreeDungeon.java
@@ -5,7 +5,7 @@ import com.telepathicgrunt.the_bumblezone.configs.BzWorldgenConfigs;
 import com.telepathicgrunt.the_bumblezone.mixin.world.WorldGenRegionAccessor;
 import com.telepathicgrunt.the_bumblezone.modinit.BzBlocks;
 import com.telepathicgrunt.the_bumblezone.modinit.BzTags;
-import com.telepathicgrunt.the_bumblezone.worldgen.features.configs.NbtFeatureConfig;
+import com.telepathicgrunt.the_bumblezone.worldgen.features.configs.TreeDungeonFeatureConfig;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.Holder;
@@ -13,30 +13,49 @@ import net.minecraft.core.Registry;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.server.level.WorldGenRegion;
 import net.minecraft.world.level.StructureManager;
+import net.minecraft.world.level.WorldGenLevel;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.PinkPetalsBlock;
+import net.minecraft.world.level.block.SaplingBlock;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.levelgen.feature.ConfiguredFeature;
 import net.minecraft.world.level.levelgen.feature.FeaturePlaceContext;
 import net.minecraft.world.level.levelgen.structure.Structure;
 import net.minecraft.world.level.levelgen.structure.StructureStart;
 
+import java.util.Optional;
 
-public class BeeDungeon extends NbtFeature<NbtFeatureConfig> {
 
-    public BeeDungeon(Codec<NbtFeatureConfig> configFactory) {
+public class TreeDungeon extends NbtFeature<TreeDungeonFeatureConfig> {
+
+    public TreeDungeon(Codec<TreeDungeonFeatureConfig> configFactory) {
         super(configFactory);
     }
 
     @Override
-    public boolean place(FeaturePlaceContext<NbtFeatureConfig> context) {
+    public boolean place(FeaturePlaceContext<TreeDungeonFeatureConfig> context) {
         //affect rarity
-        if (BzWorldgenConfigs.beeDungeonRarity >= 1000 ||
-            context.random().nextInt(BzWorldgenConfigs.beeDungeonRarity) != 0) {
+        if (BzWorldgenConfigs.treeDungeonRarity >= 1000 ||
+            context.random().nextInt(BzWorldgenConfigs.treeDungeonRarity) != 0) {
             return false;
         }
 
         if (isValidDungeonSpot(context)) {
             // generate dungeon
             super.place(context);
+
+            WorldGenLevel level = context.level();
+            BlockPos center = context.origin().above(context.config().structureYOffset + 3);
+
+            level.setBlock(center, Blocks.CAVE_AIR.defaultBlockState(), 3);
+            level.setBlock(center.above(), Blocks.CAVE_AIR.defaultBlockState(), 3);
+
+            Optional<ConfiguredFeature<?, ?>> configuredFeature = level.registryAccess()
+                    .registryOrThrow(Registries.CONFIGURED_FEATURE)
+                    .getOptional(context.config().treeConfiguredFeature);
+
+            configuredFeature.ifPresent(cf -> cf.place(level, context.chunkGenerator(), context.random(), center));
         }
 
         return true;